Allow placing text read from STDIN, as well as the base, at specified locations.
- The base location can be specified via the new -y ROW and -x COL options.
- Added a new -T ROW,COL option such that when it's used, texts are
  slurped from STDIN and then displayed at the location specified.

This enhancement allows one to format a chunk of text however he/she
likes and then feed it to cbonsai with -T ROW,COL to position it.
Together with -x COL and/or -y ROW for controlling the location of the
bonsai, one can now show any text along with a bonsai in the terminal!